Tests for push
This modifies the test script to create two repos in the temp dir: an
"origin" (bare repo) and a "local" clone of it.

The tests work in the `local` repo as usual, and we examine the `git log` from
the origin repo after the tests have run, allowing us to see what refs
have been pushed.