Add a way to check for missing tools
This is quite a major change for the tests. We're isolating them to only
have access to /bin and /usr/bin in the PATH, and everything else we add
as wrapper scripts in a $workdir/bin folder delegating to your local
install of the tool.

This allows us to remove the tool from $workdir/bin and create the
scenario where you don't have that tool installed.

We've also added the facility to start a manul test session in a
particular fixuture, with all the same git repos, isolated PATH setup
etc as you'd get in a test. To start a manual test session, pass the
path to the fixture to ./polyglot-release-test