go: do not change module name
Go modules are not only used externally, they are used internally as well.
This means that when changed in a release they effectively break the released
package.

It is not feasible to change all module names with a script, and generally runs
against the preferences of the go community. Unfortunately this does mean that
we'll have to keep an eye out for breaking changes and change the module name
accordingly.