Check if release exists against remote
It was not possible to make a first release for any project as
polyglot-release would check if the git tags were fetched by checking
if the latest release from the changelog was present.

This served to abort the release if the tag already existed.

This didn't account for:

a) A repository without releases. When checking the latest release, the
   `changelog` tool would error. So polyglot release would abort.

b) Releases can have multiple tags. For example releases that include go
   are tagged with v0.0.0 and go/v0.0.0.

By checking if the tag exists on the remote we avoid both these
problems.